mirage function is now an object constructorand all the stuff was abstracted out need to add unopinionated method to start webrtc process but retain video and canvas accessibility
